refactor: add some life to fav page

+++ src/components/layout/fav/FavCard.tsx
@@ -12,6 +12,7 @@ import {
   isInCartSelector,
   removeItem,
 } from "@/frontend/redux/Slice/CartSlice";
+import { useRouter } from "next/navigation";
 
 interface FavCardProps {
   product: IProduct;
@@ -19,9 +20,10 @@ interface FavCardProps {
 
 const FavCard = ({ product }: FavCardProps) => {
   const t = useTranslations("Favorites.card");
+  const router = useRouter();
   const { view } = useAppSelector((state: RootState) => state.fav);
   const isInCart = useAppSelector((state: RootState) =>
-    isInCartSelector(state, product.id)
+    isInCartSelector(state, product.id),
   );
   const dispatch = useAppDispatch();
   const handleRemoveFromFav = () => {
@@ -30,6 +32,7 @@ const FavCard = ({ product }: FavCardProps) => {
   };
 
   const handleAddToCart = () => {
+    if (!product.quantity) return;
     if (isInCart) {
       dispatch(removeItem(product.id));
       toast.success(t("removedFromCart"));
@@ -39,7 +42,7 @@ const FavCard = ({ product }: FavCardProps) => {
           ...product,
           quantity: 1,
           maxQuantity: product.quantity || 999,
-        })
+        }),
       );
       toast.success(t("addedToCart"));
     }
@@ -47,9 +50,10 @@ const FavCard = ({ product }: FavCardProps) => {
 
   return (
     <div
+      onClick={() => router.push(`/store/${product.id}`)}
       className={`${
         view === "grid" ? "flex-col" : "flex-row items-center gap-4 h-75 p-2"
-      } flex shadow-md rounded-lg dark:bg-primary/10 bg-background/50 relative`}
+      } flex cursor-pointer shadow-md rounded-lg dark:bg-primary/10 bg-background/50 relative`}
     >
       {/* img */}
       <div className={`${view === "grid" ? "w-full h-62.5" : "w-1/3 h-full"}`}>
@@ -62,18 +66,26 @@ const FavCard = ({ product }: FavCardProps) => {
         />
       </div>
       <div className="p-3 capitalize flex-1 flex flex-col justify-between gap-2">
-        <h3 className="text-foreground text-md md:text-lg">
-          {product.productName}
-        </h3>
+        <div className="flex justify-between text-foreground text-md md:text-lg">
+          <span>{product.productName}</span>
+          <span>{product.productPrice} EGP</span>
+        </div>
         {view === "horizontal" && (
-          <>
-            <p className="text-primary">{product.productSubtitle}</p>
-          </>
+          <p className="text-primary">{product.productSubtitle}</p>
         )}
-        {/* <p className="text-primary">${product.price.toFixed(2)}</p> */}
-        <div className="flex justify-evenly mt-4 gap-5">
-          <button className="myBtnPrimary" onClick={handleAddToCart}>
-            {isInCart ? t("removeFromCart") : t("addToCart")}
+        <div className="flex justify-evenly gap-5">
+          <button
+            className={`myBtnPrimary ${
+              !product.quantity ? "cursor-not-allowed! opacity-50" : ""
+            }`}
+            onClick={handleAddToCart}
+            disabled={!product.quantity}
+          >
+            {product.quantity
+              ? isInCart
+                ? t("removeFromCart")
+                : t("addToCart")
+              : t("outOfStock")}
           </button>
         </div>
       </div>
